优化jobhandler
