+++ exchangis-job/exchangis-job-builder/src/main/java/com/webank/wedatasphere/exchangis/job/datax/handler/HiveJobHandler.java
@@ -1,34 +1,18 @@
 package com.webank.wedatasphere.exchangis.job.datax.handler;
 
-import com.webank.wedatasphere.exchangis.job.datax.reader.HiveReader;
 import com.webank.wedatasphere.exchangis.job.datax.reader.Reader;
-import com.webank.wedatasphere.exchangis.job.datax.writer.HiveWriter;
 import com.webank.wedatasphere.exchangis.job.datax.writer.Writer;
 import com.webank.wedatasphere.exchangis.job.domain.ExchangisSubJob;
 
-import java.util.HashMap;
-import java.util.Map;
-
 public class HiveJobHandler extends DataxJobHandler {
-    @Override
-    public Reader handleReader(ExchangisSubJob job, Long jobId) {
-        HiveReader reader = new HiveReader();
-        Map params = new HashMap<String, String>();
+
+    public void handleReader(ExchangisSubJob subjob, Long jobId, Reader reader) {
         reader.setName("hivereader");
-        params.put("username", "root");
-        params.put("password", "s34uhWG0Lv*");
-        reader.setParameter(params);
-        return reader;
+        super.handleReader(subjob, jobId, reader);
     }
 
-    @Override
-    public Writer handleWriter(ExchangisSubJob job, Long jobId) {
-        HiveWriter writer = new HiveWriter();
-        Map params = new HashMap<String, String>();
+    public void handleWriter(ExchangisSubJob subjob, Long jobId, Writer writer) {
         writer.setName("hivewriter");
-        params.put("username", "root");
-        params.put("password", "s34uhWG0Lv*");
-        writer.setParameter(params);
-        return writer;
+        super.handleWriter(subjob, jobId, writer);
     }
 }
