+++ exchangis-job/exchangis-job-builder/src/main/java/com/webank/wedatasphere/exchangis/job/datax/reader/HiveReader.java
@@ -1,16 +1,5 @@
 package com.webank.wedatasphere.exchangis.job.datax.reader;
 
-import java.util.Map;
-
 public class HiveReader extends Reader {
 
-    private Map parameter;
-
-    public Map getParameter() {
-        return parameter;
-    }
-
-    public void setParameter(Map parameter) {
-        this.parameter = parameter;
-    }
 }
