+++ exchangis-job/exchangis-job-builder/src/main/java/com/webank/wedatasphere/exchangis/job/datax/reader/MysqlReader.java
@@ -1,16 +1,6 @@
 package com.webank.wedatasphere.exchangis.job.datax.reader;
 
-import java.util.Map;
-
 public class MysqlReader extends Reader {
 
-    private Map parameter;
 
-    public Map getParameter() {
-        return parameter;
-    }
-
-    public void setParameter(Map parameter) {
-        this.parameter = parameter;
-    }
 }
