+++ exchangis-job/exchangis-job-builder/src/main/java/com/webank/wedatasphere/exchangis/job/datax/writer/HiveWriter.java
@@ -1,16 +1,5 @@
 package com.webank.wedatasphere.exchangis.job.datax.writer;
 
-import java.util.Map;
-
 public class HiveWriter extends Writer {
 
-    private Map parameter;
-
-    public Map getParameter() {
-        return parameter;
-    }
-
-    public void setParameter(Map parameter) {
-        this.parameter = parameter;
-    }
 }
