+++ exchangis-job/exchangis-job-builder/src/main/java/com/webank/wedatasphere/exchangis/job/datax/writer/Writer.java
@@ -2,6 +2,16 @@ package com.webank.wedatasphere.exchangis.job.datax.writer;
 
 import com.webank.wedatasphere.exchangis.job.datax.domain.Content;
 
-public class Writer extends Content {
+import java.util.Map;
 
+public class Writer extends Content {
+    private Map parameter;
+
+    public Map getParameter() {
+        return parameter;
+    }
+
+    public void setParameter(Map parameter) {
+        this.parameter = parameter;
+    }
 }
