+++ exchangis-job/exchangis-job-builder/src/main/java/com/webank/wedatasphere/exchangis/job/handler/JobHandler.java
@@ -1,13 +1,12 @@
 package com.webank.wedatasphere.exchangis.job.handler;
 
-import com.webank.wedatasphere.exchangis.datasource.core.exception.ExchangisDataSourceException;
 import com.webank.wedatasphere.exchangis.job.datax.reader.Reader;
 import com.webank.wedatasphere.exchangis.job.datax.writer.Writer;
 import com.webank.wedatasphere.exchangis.job.domain.ExchangisSubJob;
 
 public interface JobHandler {
 
-    Reader handleReader(ExchangisSubJob job, Long jobId) throws ExchangisDataSourceException;
+    void handleReader(ExchangisSubJob job, Long jobId, Reader reader);
 
-    Writer handleWriter(ExchangisSubJob job, Long jobId);
+    void handleWriter(ExchangisSubJob job, Long jobId, Writer writer);
 }
